+++ features/openMultiplyVault/openMultiplyQuote.ts
@@ -60,7 +60,7 @@ export function applyQuote(
       map(quoteToChange),
     )
   }
-  return EMPTY
+  // return EMPTY
   return state.quote === undefined ? EMPTY : of({ kind: 'quoteReset' })
 }
